refactored stats bar
