+++ src/components/pokemon-information/stat-bars/style.ts
@@ -3,6 +3,15 @@ import { fontBgColour } from "../../../utils/colours";
 const STAT_BAR_HEIGHT = "15px";
 const STAT_BAR_RADIUS = "60px";
 
+export const statsContainer = {
+  display: "flex",
+  flexDirection: "column",
+  gap: "7px",
+  justifyContent: "center",
+  padding: "0 12px",
+  width: "100%",
+};
+
 export const statContainer = {
   display: "flex",
   width: "100%",
@@ -41,4 +50,14 @@ export const statTitle = {
   position: "absolute",
   left: "5px",
   top: "0",
+};
+
+export const statTotalContainer = {
+  display: "flex",
+  justifyContent: "space-between",
+  bgcolor: "#f0f3f8",
+  borderRadius: "5px",
+  p: "2px 15px",
+  m: "5px auto",
+  width: "80%",
 };
